% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{adjustangle}
\alias{adjustangle}
\title{Calculate the angle inside the triangle made by the fish's escape path,
   the line perpendicular to the net boundary, and the net boundary itself.
   This is necessary for escape time calculations.}
\usage{
adjustangle(anglevec)
}
\description{
@param anglevec A vector of angles in radians
 @return A vector of angles (in radians) relative to the escape path
 @exmaples
 adjustangle(3*pi/2)
}
