% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{escfish}
\alias{escfish}
\title{Calculates which fish escape the net.}
\usage{
escfish(fish, pop, swim.fail = 0.4)
}
\arguments{
\item{fish}{A data.frame of fish information, created by mkfish}

\item{pop}{A data.frame of tow information, created by mktow}

\item{swim.fail}{Numeric [0, 1]. Proportion of fish that fail to swim.}
}
\value{
Returns the original fish data.frame with new columns for
pop values, fish position, and caught status
}
\description{
Calculates which fish escape the net.
}
\examples{
escfish(fish3, tow3)
}
