% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{mkfish}
\alias{mkfish}
\title{Create a dataset of fish information.}
\usage{
mkfish(n.pops, n.fish, width.path)
}
\arguments{
\item{n.pops}{Numeric. The number of tows to complete (or fish populations to simulate)}

\item{n.fish}{Numeric. Tumber of fish to place in the path of the gear for each population.}

\item{width.path}{Numeric. The width of the path of the gear (in cm).}
}
\value{
A data.frame with one row for each fish and columns
width.path = width of the gear/net in cm,
pop = a population of fish for each tow of the net,
distance = distance in cm from edge of net,
angle = angle in radians
}
\description{
Create a dataset of fish information.
}
\examples{
mkfish(n.pops = 1, n.fish = 100, width.path = 100)
}
