% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{sumcatch}
\alias{sumcatch}
\title{Summarizes the proportion of fish caught.}
\usage{
sumcatch(fish, pop, n.fish)
}
\arguments{
\item{fish}{A data.frame of fish information, created by mkfish and modifed by escfish.}

\item{pop}{A data.frame of tow information, created by mktow.}

\item{n.fish}{Numeric. Tumber of fish to place in the path of the gear for each population.}
}
\value{
Returns a data.frame of pop with catch and catch.p appended to it
}
\description{
Summarizes the proportion of fish caught.
}
\examples{
escfish(fish3, tow3, n.fish = 1000)
}
