% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tows.R
\name{vel}
\alias{vel}
\title{Convenience function that calculates tow velocity from volume data.}
\usage{
vel(x)
}
\arguments{
\item{x}{Numeric. A tow volume (in units?).}
}
\value{
A tow velocity, assuming a 12-minute tow for the IEP FMWT net dimensions.
}
\description{
Convenience function that calculates tow velocity from volume data.
}
\examples{
vel(100)
}
