% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tows.R
\name{mktow}
\alias{mktow}
\title{Creates a dataset of tow information.}
\usage{
mktow(n.pops, net.vel, n.fish, rxn.dist)
}
\arguments{
\item{n.pops}{Numeric. The number of tows to complete (or fish populations to simulate)}

\item{net.vel}{Numeric. The velocity of the gear moving through the water.}

\item{n.fish}{Numeric. Tumber of fish to place in the path of the gear for each population.}

\item{rxn.dist}{Numeric. The reaction distance of the fish.}
}
\value{
A data.frame with one row for each population or tow. Columns
n.pops = number of tows to complete (or fish populations to simulate)
net.vel = velocity for the tows; constant
n.fish = number of fish to place in the path of the gear
rxn.dist = reaction distance in cm
}
\description{
Creates a dataset of tow information.
}
